feature: delete and optionally recreate empty chats on leaving

+++ src/test/java/de/caritas/cob/userservice/api/adapters/web/controller/UserControllerE2EIT.java
@@ -20,6 +20,7 @@ import static org.hamcrest.Matchers.nullValue;
 import static org.hamcrest.Matchers.startsWith;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -51,6 +52,7 @@ import de.caritas.cob.userservice.api.actions.chat.StopChatActionCommand;
 import de.caritas.cob.userservice.api.adapters.keycloak.dto.KeycloakLoginResponseDTO;
 import de.caritas.cob.userservice.api.adapters.rocketchat.RocketChatCredentialsProvider;
 import de.caritas.cob.userservice.api.adapters.rocketchat.dto.StandardResponseDTO;
+import de.caritas.cob.userservice.api.adapters.rocketchat.dto.group.GroupDeleteResponseDTO;
 import de.caritas.cob.userservice.api.adapters.rocketchat.dto.group.GroupMemberDTO;
 import de.caritas.cob.userservice.api.adapters.rocketchat.dto.group.GroupMemberResponseDTO;
 import de.caritas.cob.userservice.api.adapters.rocketchat.dto.message.MessageResponse;
@@ -83,6 +85,7 @@ import de.caritas.cob.userservice.api.config.auth.IdentityConfig;
 import de.caritas.cob.userservice.api.config.auth.UserRole;
 import de.caritas.cob.userservice.api.exception.rocketchat.RocketChatUserNotInitializedException;
 import de.caritas.cob.userservice.api.helper.AuthenticatedUser;
+import de.caritas.cob.userservice.api.helper.CustomLocalDateTime;
 import de.caritas.cob.userservice.api.helper.UsernameTranscoder;
 import de.caritas.cob.userservice.api.model.Chat;
 import de.caritas.cob.userservice.api.model.ChatAgency;
@@ -113,8 +116,10 @@ import java.net.URI;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -296,11 +301,17 @@ class UserControllerE2EIT {
 
   private Chat chat;
 
+  private ChatAgency chatAgency;
+
+  private UserAgency userAgency;
+
   private PasswordDTO passwordDto;
 
   private DeleteUserAccountDTO deleteUserAccountDto;
 
   private UserInfoResponseDTO userInfoResponse;
+
+  private GroupDeleteResponseDTO groupDeleteResponse;
 
   private GroupMemberResponseDTO groupMemberResponseDTO;
 
@@ -339,15 +350,24 @@ class UserControllerE2EIT {
     email = null;
     patchUserDTO = null;
     userDTO = null;
-    if (nonNull(chat)) {
+    if (nonNull(chat) && chatRepository.existsById(chat.getId())) {
       chatRepository.deleteById(chat.getId());
-      chat = null;
     }
+    chat = null;
+    if (nonNull(chatAgency) && chatAgencyRepository.existsById(chatAgency.getId())) {
+      chatAgencyRepository.deleteById(chatAgency.getId());
+    }
+    chatAgency = null;
+    if (nonNull(userAgency) && userAgencyRepository.existsById(userAgency.getId())) {
+      userAgencyRepository.deleteById(userAgency.getId());
+    }
+    userAgency = null;
     videoChatConfig.setE2eEncryptionEnabled(false);
     passwordDto = null;
     deleteUserAccountDto = null;
     userInfoResponse = null;
     subscriptionsGetResponse = null;
+    groupDeleteResponse = null;
     identityConfig.setDisplayNameAllowedForConsultants(false);
     infix = null;
     monitoringDTO = null;
@@ -2419,6 +2439,156 @@ class UserControllerE2EIT {
   }
 
   @Test
+  @WithMockUser(authorities = AuthorityValue.USER_DEFAULT)
+  void leaveChatShouldReturnOkAndDeleteOneTimeChatAndChatAgencyIfLastUser(
+      CapturedOutput logOutput) throws Exception {
+    givenAValidUser(true);
+    givenAValidConsultant();
+    givenAValidChat(false);
+    givenAValidRocketChatSystemUser();
+    givenValidRocketChatTechUserResponse();
+    givenAnOnlyTechUserRocketChatGroupMemberResponse(chat.getGroupId());
+    givenAValidRocketChatInfoUserResponse();
+    givenAValidRocketChatGroupDeleteResponse();
+
+    mockMvc.perform(
+            put("/users/chat/{chatId}/leave", chat.getId())
+                .cookie(CSRF_COOKIE)
+                .header(CSRF_HEADER, CSRF_VALUE)
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk());
+
+    assertFalse(chatRepository.existsById(chat.getId()));
+    assertFalse(chatAgencyRepository.existsById(chatAgency.getId()));
+
+    var urlSuffix = "/api/v1/groups.delete";
+    verify(restTemplate).postForObject(
+        endsWith(urlSuffix), any(HttpEntity.class), eq(GroupDeleteResponseDTO.class)
+    );
+    verifyRocketChatUserRemovedFromGroup(logOutput, chat.getGroupId(), user.getRcUserId(), 1);
+  }
+
+  @Test
+  @WithMockUser(authorities = AuthorityValue.USER_DEFAULT)
+  void leaveChatShouldReturnOkAndNotDeleteOneTimeChatOrChatAgencyIfNotLastUser(
+      CapturedOutput logOutput) throws Exception {
+    givenAValidUser(true);
+    givenAValidConsultant();
+    givenAValidChat(false);
+    givenAValidRocketChatSystemUser();
+    givenValidRocketChatTechUserResponse();
+    var chatUserId = RandomStringUtils.randomAlphanumeric(17);
+    givenAPositiveRocketChatGroupMemberResponse(chat.getGroupId(), chatUserId);
+    givenAValidRocketChatInfoUserResponse();
+
+    mockMvc.perform(
+            put("/users/chat/{chatId}/leave", chat.getId())
+                .cookie(CSRF_COOKIE)
+                .header(CSRF_HEADER, CSRF_VALUE)
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk());
+
+    assertTrue(chatRepository.existsById(chat.getId()));
+    assertTrue(chatAgencyRepository.existsById(chatAgency.getId()));
+
+    var urlSuffix = "/api/v1/groups.delete";
+    verify(restTemplate, never()).postForObject(
+        endsWith(urlSuffix), any(HttpEntity.class), eq(GroupDeleteResponseDTO.class)
+    );
+    verifyRocketChatUserRemovedFromGroup(logOutput, chat.getGroupId(), user.getRcUserId(), 1);
+  }
+
+  @Test
+  @WithMockUser(authorities = AuthorityValue.USER_DEFAULT)
+  void leaveChatShouldReturnOkAndDeleteRepetitiveChatAndChatAgencyAsWellAsRecreateIfLastUser(
+      CapturedOutput logOutput) throws Exception {
+    givenAValidUser(true);
+    givenAValidConsultant();
+    givenAValidChat(true);
+    givenAValidRocketChatSystemUser();
+    givenValidRocketChatTechUserResponse();
+    givenAnOnlyTechUserRocketChatGroupMemberResponse(chat.getGroupId());
+    givenAValidRocketChatInfoUserResponse();
+    givenAValidRocketChatGroupDeleteResponse();
+
+    var allChatsBefore = StreamSupport.stream(chatRepository.findAll().spliterator(), false)
+        .collect(Collectors.toSet());
+
+    mockMvc.perform(
+            put("/users/chat/{chatId}/leave", chat.getId())
+                .cookie(CSRF_COOKIE)
+                .header(CSRF_HEADER, CSRF_VALUE)
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk());
+
+    var allChatsAfter = StreamSupport.stream(chatRepository.findAll().spliterator(), false)
+        .sorted(Comparator.comparing(Chat::getUpdateDate).reversed())
+        .collect(Collectors.toList());
+    assertEquals(allChatsBefore.size(), allChatsAfter.size());
+    var recreatedChat = allChatsAfter.get(0);
+    var keptChats = allChatsAfter.subList(1, allChatsAfter.size());
+    assertFalse(allChatsBefore.contains(recreatedChat));
+    assertTrue(allChatsBefore.containsAll(keptChats));
+
+    assertFalse(chatRepository.existsById(chat.getId()));
+    assertFalse(chatAgencyRepository.existsById(chatAgency.getId()));
+
+    var urlSuffix = "/api/v1/groups.delete";
+    verify(restTemplate).postForObject(
+        endsWith(urlSuffix), any(HttpEntity.class), eq(GroupDeleteResponseDTO.class)
+    );
+    verifyRocketChatUserRemovedFromGroup(logOutput, chat.getGroupId(), user.getRcUserId(), 1);
+
+    assertEquals(chat.getTopic(), recreatedChat.getTopic());
+    assertEquals(chat.getConsultingTypeId(), recreatedChat.getConsultingTypeId());
+    assertEquals(chat.getInitialStartDate().truncatedTo(ChronoUnit.SECONDS),
+        recreatedChat.getInitialStartDate().truncatedTo(ChronoUnit.SECONDS));
+    assertTrue(chat.getStartDate().isBefore(recreatedChat.getStartDate()));
+    assertEquals(chat.getDuration(), recreatedChat.getDuration());
+    assertEquals(chat.isRepetitive(), recreatedChat.isRepetitive());
+    assertEquals(chat.getChatInterval(), recreatedChat.getChatInterval());
+    assertFalse(recreatedChat.isActive());
+    assertEquals(chat.getMaxParticipants(), recreatedChat.getMaxParticipants());
+    assertNotEquals(chat.getGroupId(), recreatedChat.getGroupId());
+    assertNotNull(recreatedChat.getGroupId());
+    assertEquals(chat.getChatOwner(), recreatedChat.getChatOwner());
+    assertTrue(recreatedChat.getChatAgencies().size() > 0);
+    assertTrue(chat.getUpdateDate().isBefore(recreatedChat.getUpdateDate()));
+
+    chat = recreatedChat;
+  }
+
+  @Test
+  @WithMockUser(authorities = AuthorityValue.USER_DEFAULT)
+  void leaveChatShouldReturnOkAndNotDeleteRepetitiveChatOrChatAgencyIfNotLastUser(
+      CapturedOutput logOutput) throws Exception {
+    givenAValidUser(true);
+    givenAValidConsultant();
+    givenAValidChat(true);
+    givenAValidRocketChatSystemUser();
+    givenValidRocketChatTechUserResponse();
+    var chatUserId = RandomStringUtils.randomAlphanumeric(17);
+    givenAPositiveRocketChatGroupMemberResponse(chat.getGroupId(), chatUserId);
+    givenAValidRocketChatInfoUserResponse();
+
+    mockMvc.perform(
+            put("/users/chat/{chatId}/leave", chat.getId())
+                .cookie(CSRF_COOKIE)
+                .header(CSRF_HEADER, CSRF_VALUE)
+                .accept(MediaType.APPLICATION_JSON))
+        .andExpect(status().isOk());
+
+    assertTrue(chatRepository.existsById(chat.getId()));
+    assertTrue(chatAgencyRepository.existsById(chatAgency.getId()));
+
+    var urlSuffix = "/api/v1/groups.delete";
+    verify(restTemplate, never()).postForObject(
+        endsWith(urlSuffix), any(HttpEntity.class), eq(GroupDeleteResponseDTO.class)
+    );
+    verifyRocketChatUserRemovedFromGroup(logOutput, chat.getGroupId(), user.getRcUserId(), 1);
+  }
+
+  @Test
   @WithMockUser(authorities = AuthorityValue.STOP_CHAT)
   void stopChatShouldReturnOkIfUsersAreBanned() throws Exception {
     givenAValidUser();
@@ -3688,11 +3858,40 @@ class UserControllerE2EIT {
     ).thenReturn(ResponseEntity.ok(userInfoResponse));
   }
 
+  private void givenAValidRocketChatGroupDeleteResponse() {
+    groupDeleteResponse = new GroupDeleteResponseDTO();
+    groupDeleteResponse.setSuccess(true);
+
+    var urlSuffix = "/api/v1/groups.delete";
+    when(restTemplate.postForObject(endsWith(urlSuffix), any(HttpEntity.class),
+        eq(GroupDeleteResponseDTO.class))
+    ).thenReturn(groupDeleteResponse);
+  }
+
   private void givenAPositiveRocketChatGroupMemberResponse(String chatId, String chatUserId) {
     groupMemberResponseDTO = new GroupMemberResponseDTO();
     groupMemberResponseDTO.setSuccess(true);
 
     var groupMember = easyRandom.nextObject(GroupMemberDTO.class);
+    if (nonNull(chatUserId)) {
+      groupMember.set_id(chatUserId);
+    }
+    GroupMemberDTO[] groupMembers = {groupMember};
+    groupMemberResponseDTO.setMembers(groupMembers);
+
+    var urlSuffix = "/api/v1/groups.members?roomId=" + chatId + "&count=0";
+    when(restTemplate.exchange(
+        endsWith(urlSuffix), eq(HttpMethod.GET),
+        any(HttpEntity.class), eq(GroupMemberResponseDTO.class))
+    ).thenReturn(ResponseEntity.ok(groupMemberResponseDTO));
+  }
+
+  private void givenAnOnlyTechUserRocketChatGroupMemberResponse(String chatId) {
+    groupMemberResponseDTO = new GroupMemberResponseDTO();
+    groupMemberResponseDTO.setSuccess(true);
+
+    var groupMember = easyRandom.nextObject(GroupMemberDTO.class);
+    var chatUserId = RC_CREDENTIALS_TECHNICAL_A.getRocketChatUserId();
     if (nonNull(chatUserId)) {
       groupMember.set_id(chatUserId);
     }
@@ -3941,16 +4140,17 @@ class UserControllerE2EIT {
     chat.setConsultingTypeId(easyRandom.nextInt(128));
     chat.setDuration(easyRandom.nextInt(32768));
     chat.setMaxParticipants(easyRandom.nextInt(128));
+    chat.setUpdateDate(CustomLocalDateTime.nowInUtc());
     chatRepository.save(chat);
 
     var agencyId = consultant.getConsultantAgencies().iterator().next().getAgencyId();
-    var chatAgency = new ChatAgency();
+    chatAgency = new ChatAgency();
     chatAgency.setChat(chat);
     chatAgency.setAgencyId(agencyId);
     chatAgencyRepository.save(chatAgency);
 
     if (nonNull(user)) {
-      var userAgency = new UserAgency();
+      userAgency = new UserAgency();
       userAgency.setUser(user);
       userAgency.setAgencyId(agencyId);
       user.getUserAgencies().add(userAgency);
